+++ packages/static/src/lib/paths.ts
@@ -1,12 +1,12 @@
-import path from "node:path"
+import path from 'node:path'
 
 // Used as an optimization to avoid dual lookups for missing assets
 const assetExtensionRegExp = /\.(html?|png|jpg|js|css|svg|gif|ico|woff|woff2)$/
 
-export const getFilePathsForURL = (pathname: string, baseDirectory = "") => {
+export const getFilePathsForURL = (pathname: string, baseDirectory = '') => {
   const urlVariations = getURLVariations(pathname)
-  const possiblePaths = urlVariations.map(urlVariation => {
-    const parts = urlVariation.split("/").filter(Boolean) as string[]
+  const possiblePaths = urlVariations.map((urlVariation) => {
+    const parts = urlVariation.split('/').filter(Boolean)
 
     return path.resolve.apply(null, [baseDirectory, ...parts])
   })
@@ -16,7 +16,7 @@ export const getFilePathsForURL = (pathname: string, baseDirectory = "") => {
 
 export const getURLVariations = (pathname: string) => {
   const paths: string[] = []
-  
+
   if (pathname[pathname.length - 1] === '/') {
     const end = pathname.length - 1
 
@@ -29,7 +29,7 @@ export const getURLVariations = (pathname: string) => {
     paths.push(`${pathname}.html`, `${pathname}.htm`, `${pathname}/index.html`, `${pathname}/index.htm`)
   }
 
-  if (!(paths.includes(pathname))) {
+  if (!paths.includes(pathname)) {
     return [pathname, ...paths]
   }